+++ R/bread.R
@@ -20,8 +20,8 @@ bread.mlm <- function(x, ...)
   cf <- coef(x)
   rval <- summary.lm(x)
   rval <- kronecker(
-    structure(diag(ncol(cf)), .Dimnames = rep.int(list(colnames(cf)), 2L)),
-    structure(rval$cov.unscaled,  .Dimnames = rep.int(list(rownames(cf)), 2L)) * as.vector(sum(rval$df[1L:2L])),
+    structure(diag(ncol(cf)), dimnames = rep.int(list(colnames(cf)), 2L)),
+    structure(rval$cov.unscaled,  dimnames = rep.int(list(rownames(cf)), 2L)) * as.vector(sum(rval$df[1L:2L])),
     make.dimnames = TRUE
   )
   return(rval)